refactor: consolidate compiler resolution
